+++ silhouette/app/com/mohiva/play/silhouette/api/Silhouette.scala
@@ -1,4 +1,18 @@
-
+/**
+ * Copyright 2015 Mohiva Organisation (license at mohiva dot com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.mohiva.play.silhouette.api
 
 import javax.inject.Inject
@@ -25,6 +39,11 @@ trait Silhouette[E <: Env] {
   val securedAction: SecuredAction
 
   /**
+   * The unsecured action stack.
+   */
+  val unsecuredAction: UnsecuredAction
+
+  /**
    * The user aware action stack.
    */
   val userAwareAction: UserAwareAction
@@ -42,6 +61,20 @@ trait Silhouette[E <: Env] {
    * @return The secured request handler implementation.
    */
   def SecuredRequestHandler: SecuredRequestHandlerBuilder[E] = securedAction.requestHandler(env)
+
+  /**
+   * Provides the unsecured action implementation.
+   *
+   * @return The unsecured action implementation.
+   */
+  def UnsecuredAction: UnsecuredActionBuilder[E] = unsecuredAction(env)
+
+  /**
+   * Provides the unsecured request handler implementation.
+   *
+   * @return The unsecured request handler implementation.
+   */
+  def UnsecuredRequestHandler: UnsecuredRequestHandlerBuilder[E] = unsecuredAction.requestHandler(env)
 
   /**
    * Provides the user-aware action implementation.
@@ -69,5 +102,6 @@ trait Silhouette[E <: Env] {
 class SilhouetteProvider[E <: Env] @Inject() (
   val env: Environment[E],
   val securedAction: SecuredAction,
+  val unsecuredAction: UnsecuredAction,
   val userAwareAction: UserAwareAction)
   extends Silhouette[E]